LC-put-marbles-in-bags [Runtime: 34 ms; Memory: 63.5 MB]

+++ heap-priority-queue/put-marbles-in-bags.java
@@ -3,7 +3,7 @@ class Solution {
         int n = weights.length;
         long pre[] = new long[n-1];
         for(int i = 0; i < n-1; i++) {
-            pre[i] = (long)(weights[i] + weights[i+1]);
+            pre[i] = weights[i] + weights[i+1];
         }
         Arrays.sort(pre);
         long ans = 0l;
